Move definitions, errors, and converters into their own modules. Move definitions, errors and converters tests to their own modules. (tests pass locally)
