+++ pint/unit.py
@@ -25,310 +25,13 @@ from tokenize import untokenize, NUMBER, STRING, NAME, OP
 from .context import Context, ContextChain
 from .util import (logger, pi_theorem, solve_dependencies, ParserHelper,
                    string_preprocessor, find_connected_nodes,
-                   find_shortest_path, UnitsContainer)
+                   find_shortest_path, UnitsContainer, _is_dim)
 from .compat import tokenizer, string_types
-
-
-class DefinitionSyntaxError(ValueError):
-    """Raised when a textual definition has a syntax error.
-    """
-
-    def __init__(self, msg, filename=None, lineno=None):
-        super(ValueError, self).__init__()
-        self.msg = msg
-        self.filename = None
-        self.lineno = None
-
-    def __str__(self):
-        return "While opening {0}, in line {1}: ".format(self.filename, self.lineno) + self.msg
-
-
-class RedefinitionError(ValueError):
-    """Raised when a unit or prefix is redefined.
-    """
-
-    def __init__(self, name, definition_type):
-        super(ValueError, self).__init__()
-        self.name = name
-        self.definition_type = definition_type
-        self.filename = None
-        self.lineno = None
-
-    def __str__(self):
-        msg = "cannot redefine '{0}' ({1})".format(self.name, self.definition_type)
-        if self.filename:
-            return "While opening {0}, in line {1}: ".format(self.filename, self.lineno) + msg
-        return msg
-
-
-class UndefinedUnitError(ValueError):
-    """Raised when the units are not defined in the unit registry.
-    """
-
-    def __init__(self, unit_names):
-        super(ValueError, self).__init__()
-        self.unit_names = unit_names
-
-    def __str__(self):
-        if isinstance(self.unit_names, string_types):
-            return "'{0}' is not defined in the unit registry".format(self.unit_names)
-        elif isinstance(self.unit_names, (list, tuple)) and len(self.unit_names) == 1:
-            return "'{0}' is not defined in the unit registry".format(self.unit_names[0])
-        elif isinstance(self.unit_names, set) and len(self.unit_names) == 1:
-            uname = list(self.unit_names)[0]
-            return "'{0}' is not defined in the unit registry".format(uname)
-        else:
-            return '{0} are not defined in the unit registry'.format(self.unit_names)
-
-
-class DimensionalityError(ValueError):
-    """Raised when trying to convert between incompatible units.
-    """
-
-    def __init__(self, units1, units2, dim1=None, dim2=None, extra_msg=''):
-        super(DimensionalityError, self).__init__()
-        self.units1 = units1
-        self.units2 = units2
-        self.dim1 = dim1
-        self.dim2 = dim2
-        self.extra_msg = extra_msg
-
-    def __str__(self):
-        if self.dim1 or self.dim2:
-            dim1 = ' ({0})'.format(self.dim1)
-            dim2 = ' ({0})'.format(self.dim2)
-        else:
-            dim1 = ''
-            dim2 = ''
-
-        msg = "Cannot convert from '{0}'{1} to '{2}'{3}" + self.extra_msg
-
-        return msg.format(self.units1, dim1, self.units2, dim2)
-
-
-class OffsetUnitCalculusError(ValueError):
-    """Raised on ambiguous operations with offset units.
-    """
-    def __init__(self, units1, units2='', extra_msg=''):
-        super(ValueError, self).__init__()
-        self.units1 = units1
-        self.units2 = units2
-        self.extra_msg = extra_msg
-
-    def __str__(self):
-        msg = ("Ambiguous operation with offset unit (%s)." %
-               ', '.join(['%s' % u for u in [self.units1, self.units2] if u])
-               + self.extra_msg)
-        return msg.format(self.units1, self.units2)
-
-
-class Converter(object):
-    """Base class for value converters.
-    """
-
-    is_multiplicative = True
-
-    def to_reference(self, value, inplace=False):
-        return value
-
-    def from_reference(self, value, inplace=False):
-        return value
-
-
-class ScaleConverter(Converter):
-    """A linear transformation
-    """
-
-    is_multiplicative = True
-
-    def __init__(self, scale):
-        self.scale = scale
-
-    def to_reference(self, value, inplace=False):
-        if inplace:
-            value *= self.scale
-        else:
-            value = value * self.scale
-
-        return value
-
-    def from_reference(self, value, inplace=False):
-        if inplace:
-            value /= self.scale
-        else:
-            value = value / self.scale
-
-        return value
-
-
-class OffsetConverter(Converter):
-    """An affine transformation
-    """
-
-    def __init__(self, scale, offset):
-        self.scale = scale
-        self.offset = offset
-
-    @property
-    def is_multiplicative(self):
-        return self.offset == 0
-
-    def to_reference(self, value, inplace=False):
-        if inplace:
-            value *= self.scale
-            value += self.offset
-        else:
-            value = value * self.scale + self.offset
-
-        return value
-
-    def from_reference(self, value, inplace=False):
-        if inplace:
-            value -= self.offset
-            value /= self.scale
-        else:
-            value = (value - self.offset) / self.scale
-
-        return value
-
-
-class Definition(object):
-    """Base class for definitions.
-
-    :param name: name.
-    :param symbol: a short name or symbol for the definition
-    :param aliases: iterable of other names.
-    :param converter: an instance of Converter.
-    """
-
-    def __init__(self, name, symbol, aliases, converter):
-        self._name = name
-        self._symbol = symbol
-        self._aliases = aliases
-        self._converter = converter
-
-    @property
-    def is_multiplicative(self):
-        return self._converter.is_multiplicative
-
-    @classmethod
-    def from_string(cls, definition):
-        """Parse a definition
-        """
-        name, definition = definition.split('=', 1)
-        name = name.strip()
-
-        result = [res.strip() for res in definition.split('=')]
-        value, aliases = result[0], tuple(result[1:])
-        symbol, aliases = (aliases[0], aliases[1:]) if aliases else (None, aliases)
-
-        if name.startswith('['):
-            return DimensionDefinition(name, symbol, aliases, value)
-        elif name.endswith('-'):
-            name = name.rstrip('-')
-            return PrefixDefinition(name, symbol, aliases, value)
-        else:
-            return UnitDefinition(name, symbol, aliases, value)
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def symbol(self):
-        return self._symbol or self._name
-
-    @property
-    def has_symbol(self):
-        return bool(self._symbol)
-
-    @property
-    def aliases(self):
-        return self._aliases
-
-    @property
-    def converter(self):
-        return self._converter
-
-    def __str__(self):
-        return self.name
-
-
-def _is_dim(name):
-    return name[0] == '[' and name[-1] == ']'
-
-
-class PrefixDefinition(Definition):
-    """Definition of a prefix.
-    """
-
-    def __init__(self, name, symbol, aliases, converter):
-        if isinstance(converter, string_types):
-            converter = ScaleConverter(eval(converter))
-        aliases = tuple(alias.strip('-') for alias in aliases)
-        if symbol:
-            symbol = symbol.strip('-')
-        super(PrefixDefinition, self).__init__(name, symbol, aliases, converter)
-
-
-class UnitDefinition(Definition):
-    """Definition of a unit.
-
-    :param reference: Units container with reference units.
-    :param is_base: indicates if it is a base unit.
-    """
-
-    def __init__(self, name, symbol, aliases, converter,
-                 reference=None, is_base=False):
-        self.reference = reference
-        self.is_base = is_base
-        if isinstance(converter, string_types):
-            if ';' in converter:
-                [converter, modifiers] = converter.split(';', 2)
-                modifiers = dict((key.strip(), eval(value)) for key, value in
-                                 (part.split(':') for part in modifiers.split(';')))
-            else:
-                modifiers = {}
-
-            converter = ParserHelper.from_string(converter)
-            if all(_is_dim(key) for key in converter.keys()):
-                self.is_base = True
-            elif not any(_is_dim(key) for key in converter.keys()):
-                self.is_base = False
-            else:
-                raise ValueError('Cannot mix dimensions and units in the same definition. '
-                                 'Base units must be referenced only to dimensions. '
-                                 'Derived units must be referenced only to units.')
-            self.reference = UnitsContainer(converter)
-            if modifiers.get('offset', 0.) != 0.:
-                converter = OffsetConverter(converter.scale,
-                                            modifiers['offset'])
-            else:
-                converter = ScaleConverter(converter.scale)
-
-        super(UnitDefinition, self).__init__(name, symbol, aliases, converter)
-
-
-class DimensionDefinition(Definition):
-    """Definition of a dimension.
-    """
-
-    def __init__(self, name, symbol, aliases, converter,
-                 reference=None, is_base=False):
-        self.reference = reference
-        self.is_base = is_base
-        if isinstance(converter, string_types):
-            converter = ParserHelper.from_string(converter)
-            if not converter:
-                self.is_base = True
-            elif all(_is_dim(key) for key in converter.keys()):
-                self.is_base = False
-            else:
-                raise ValueError('Base dimensions must be referenced to None. '
-                                 'Derived dimensions must only be referenced to dimensions.')
-            self.reference = UnitsContainer(converter)
-
-        super(DimensionDefinition, self).__init__(name, symbol, aliases, converter=None)
+from .definitions import (Definition, UnitDefinition, PrefixDefinition,
+                          DimensionDefinition)
+from .converters import ScaleConverter
+from .errors import (DimensionalityError, UndefinedUnitError,
+                     DefinitionSyntaxError, RedefinitionError)
 
 
 class UnitRegistry(object):
